+++ CHANGELOG.md
@@ -1,15 +1,15 @@
 # Changelog
 
-## v0.2.0-dev
+## v0.2.0
 
 ### Enhancements
 
+  * add `karma-test-file-current-buffer`, which just runs `karma-start-single-run' with just the current buffer file (`buffer-file-name`).
   * `karma-mode` should also be enabled in `coffee-mode`.
 
 ## v0.1.0
 
 ### Enhancements
 
-  * implement `karma-start`, `karma-start-single-run`,
-    `karma-start-no-single-run` and `karma-run`
+  * implement `karma-start`, `karma-start-single-run`, `karma-start-no-single-run` and `karma-run`
   * add default Keymap inside `js-mode` and `js2-mode`